added images to starting position section

+++ widgets/dictionary_widget/dictionary_deletion_handler.py
@@ -2,10 +2,11 @@ import os
 import shutil
 from typing import TYPE_CHECKING
 from variation_number_fixer import VariationNumberFixer
+from widgets.dictionary_widget.delete_confirmation_dialog import DeleteConfirmationDialog
 from widgets.path_helpers.path_helpers import get_images_and_data_path
-from PyQt6.QtWidgets import QApplication
-from PyQt6.QtCore import Qt
-
+from PyQt6.QtWidgets import QApplication, QLabel, QVBoxLayout, QPushButton
+from PyQt6.QtCore import Qt, QTimer
+from PyQt6.QtWidgets import QDialog
 if TYPE_CHECKING:
     from widgets.dictionary_widget.thumbnail_box.thumbnail_box import ThumbnailBox
     from widgets.dictionary_widget.dictionary_widget import DictionaryWidget
@@ -20,36 +21,32 @@ class DictionaryDeletionHandler:
         return not any(os.scandir(folder_path))
 
     def delete_variation(self, thumbnail_box: "ThumbnailBox", index):
-        QApplication.setOverrideCursor(Qt.CursorShape.WaitCursor)
-        file_path = thumbnail_box.thumbnails.pop(index)
-        os.remove(file_path)
-        if len(thumbnail_box.thumbnails) == 0:
-            self.delete_word(thumbnail_box.word)
+        dialog = DeleteConfirmationDialog(self.dictionary_widget)
+        if dialog.exec() == QDialog.DialogCode.Accepted:
+            QApplication.setOverrideCursor(Qt.CursorShape.WaitCursor)
+            current_scroll_position = self.dictionary_widget.browser.scroll_widget.scroll_area.verticalScrollBar().value()
+            file_path = thumbnail_box.thumbnails.pop(index)
+            os.remove(file_path)
+            if len(thumbnail_box.thumbnails) == 0:
+                self.delete_word(thumbnail_box.word)
+                self.dictionary_widget.preview_area.update_thumbnails()
+                self.dictionary_widget.browser.scroll_widget.thumbnail_boxes_dict.pop(
+                    thumbnail_box.word
+                )
+            else:
+                self.delete_empty_folders(get_images_and_data_path("dictionary"))
+                thumbnail_box.current_index = 0
+                self.dictionary_widget.browser.thumbnail_box_sorter.reload_currently_displayed_filtered_sequences()
+                thumbnail_box.image_label.update_thumbnail(thumbnail_box.current_index)
+                self.dictionary_widget.preview_area.update_thumbnails(thumbnail_box.thumbnails)
+                thumbnail_box.update_thumbnails(thumbnail_box.thumbnails)
+            def restore_scroll_position():
+                self.dictionary_widget.browser.scroll_widget.scroll_area.verticalScrollBar().setValue(current_scroll_position)
+                QApplication.restoreOverrideCursor()
 
-            self.dictionary_widget.preview_area.update_thumbnails()
-            self.dictionary_widget.browser.scroll_widget.thumbnail_boxes_dict.pop(
-                thumbnail_box.word
-            )
-        else:
-            self.delete_empty_folders(get_images_and_data_path("dictionary"))
-            thumbnail_box.current_index = 0
-            # remove the deleted variation from the browser's currently_displayed_sequences
-            self.dictionary_widget.browser.thumbnail_box_sorter.reload_currently_displayed_filtered_sequences()
-            # self.dictionary_widget.browser.thumbnail_box_sorter.sort_and_display_currently_filtered_sequences_by_method(
-            #     self.dictionary_widget.main_widget.main_window.settings_manager.dictionary.get_sort_method()
-            # )
+            QApplication.processEvents()
+            QTimer.singleShot(0, restore_scroll_position)
 
-            for index, thumbnail in enumerate(thumbnail_box.thumbnails):
-                if thumbnail == file_path:
-                    thumbnail_box.current_index = index
-                    thumbnail_box.thumbnails.pop(index)
-                    break
-            thumbnail_box.image_label.update_thumbnail(thumbnail_box.current_index)
-
-            self.dictionary_widget.preview_area.update_thumbnails(
-                thumbnail_box.thumbnails
-            )
-            thumbnail_box.update_thumbnails(thumbnail_box.thumbnails)
         QApplication.restoreOverrideCursor()
 
     def delete_word(self, base_word):
@@ -65,21 +62,12 @@ class DictionaryDeletionHandler:
         shutil.rmtree(base_path)
         self.delete_empty_folders(get_images_and_data_path("dictionary"))
         self.variation_number_fixer.ensure_sequential_versions()
-        # reload currently displayed sequences
         self.dictionary_widget.browser.thumbnail_box_sorter.reload_currently_displayed_filtered_sequences()
-        # self.dictionary_widget.browser.thumbnail_box_sorter.sort_and_display_currently_filtered_sequences_by_method(
-        #     self.dictionary_widget.main_widget.main_window.settings_manager.dictionary.get_sort_method()
-        # )
 
     def delete_empty_folders(self, root_folder):
-        def is_folder_empty(folder_path):
-            return not any(os.scandir(folder_path))
-
         for root, dirs, files in os.walk(root_folder, topdown=False):
             for dir_name in dirs:
                 dir_path = os.path.join(root, dir_name)
-                if is_folder_empty(dir_path):
-                    # set it to writable
+                if self.is_folder_empty(dir_path):
                     os.chmod(dir_path, 0o777)
-                    print(f"Deleting empty folder: {dir_path}")
                     shutil.rmtree(dir_path)
